requirements.txt
openai
langchain
langchain-community
langchain-openai
langchain-chroma
langchain-core
tiktoken
faiss-cpu
chromadb
pandas
python-dotenv
